% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_explorer.R
\name{explorer_server}
\alias{explorer_server}
\title{Explorer module server}
\usage{
explorer_server(id)
}
\arguments{
\item{id}{Unique id for the module instance.}
}
\description{
Explorer module server
}
\keyword{internal}
