% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hockey_data.R
\name{hockey_data_server}
\alias{hockey_data_server}
\title{Hockey data module server}
\usage{
hockey_data_server(id)
}
\arguments{
\item{id}{Unique id for the module instance.}
}
\description{
Hockey data module server
}
\keyword{internal}
