% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_learning_plan.R
\name{learning_plan_server}
\alias{learning_plan_server}
\title{Learning plan module server}
\usage{
learning_plan_server(id)
}
\arguments{
\item{id}{Unique id for the module instance.}
}
\description{
Learning plan module server
}
\keyword{internal}
