% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_progress.R
\name{progress_server}
\alias{progress_server}
\title{Progress tracker module server}
\usage{
progress_server(id)
}
\arguments{
\item{id}{Unique id for the module instance.}
}
\description{
Progress tracker module server
}
\keyword{internal}
